Update sixelgallery.js
Should now display as many frames as get outputted from the conversion to sixel, in the case of an animated GIF. Depending on the number of frames, it could take a lot of time to convert and display each frame. For that reason, an option to display either all frames or only the first one would be nice to include at some point.