Fix for call to unzip
SixelGallery makes calls to unzip.exe in order to open ZIP files and browse their contents.

SBBS for Windows ships with unzip.exe in /sbbs/exec, so this ZIP file extraction logic expects unzip.exe to be in that path.

On *nix, however, unzip instead just exists in the system path. The ZIP file extraction fails since it cannot find unzip.exe.

A long-term solution will ultimately involve removing external calls to unzip altogether, and using the Archive class which was introduced into SBBS as of 3.19.

For now, as an inelegant workaround, we'll test for the existence of unzip.exe in /sbbs/exec before calling it, and if it does not exist, we'll simply call unzip without supplying its path, assuming the utility exists in the system path.